Update Django settings environment and header values
Two changes were made in the Django settings - the environment value for DEBUG has been converted to int before assigning it as a boolean, and a new HTTP header for secure proxy SSL has been added. Environment values are now used to get CSRF trusted origins. This leads to a more flexible and secure configuration.