Incorrect commit message: Change homepage background color in styles.css
Updated the background color from red to black (#111) in the styles.css file to enhance the website's look and feel. The modification aims to result in increased visual engagement for users.

Correct commit message:
Change homepage background color in styles.css

Updated the background color from black (#111) to green in the styles.css file. This change aims to make the website more visually engaging for users. Also, commented out the STATICFILES_STORAGE line in settings.py file of project_1 as part of the modification process.