Incorrect commit message: Change homepage background color in styles.css
Updated the background color from black (#111) to green in the styles.css file. This change aims to make the website more visually engaging for users. Also, commented out the STATICFILES_STORAGE line in settings.py file of project_1 as part of the modification process.

Correct commit message:
Change homepage background color to green in styles.css

The background color in the styles.css file changed from red to green. This modification seeks to enhance the aesthetic appeal of the website and improve user visual engagement.